+++ packages/sqorn-sql/src/clauses/delete.js
