+++ packages/sqorn-sql/src/clauses/index.js
@@ -1,18 +1,18 @@
 module.exports = {
-  wth: require('./wth'),
-  sel: require('./sel'),
-  frm: require('./frm'),
-  whr: require('./whr'),
-  grp: require('./grp'),
-  hav: require('./hav'),
-  ord: require('./ord'),
-  lim: require('./lim'),
-  off: require('./off'),
-  del: require('./del'),
-  ret: require('./ret'),
-  ins: require('./ins'),
-  val: require('./val'),
-  upd: require('./upd'),
+  wth: require('./with'),
+  select: require('./select'),
+  from: require('./from'),
+  where: require('./where'),
+  group: require('./group'),
+  having: require('./having'),
+  order: require('./order'),
+  limit: require('./limit'),
+  offset: require('./offset'),
+  del: require('./delete'),
+  returning: require('./returning'),
+  insert: require('./insert'),
+  value: require('./value'),
+  update: require('./update'),
   set: require('./set'),
   sql: require('./sql')
 }
