+++ packages/sqorn-sql/src/clauses/insert.js
